backend/requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Firebase
firebase-admin==6.4.0

# Google Gemini AI
google-generativeai==0.3.2

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
